move files to the proper places they belong
