+++ src/domain/error/MissingPeriodForDateError.ts
