+++ src/domain/types/category.ts
