+++ src/domain/types/expense.ts
