+++ src/domain/types/period.ts
